Add common scripts/utils.py script

+++ scripts/install_svc_tpl.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 set -euo pipefail
 
-helm upgrade --install $APP oci://ghcr.io/k0rdent/catalog/charts/kgst \
+if ! ./scripts/example_chart_exists.sh; then
+  helm upgrade --install $APP oci://ghcr.io/k0rdent/catalog/charts/kgst \
   -n kcm-system \
   -f apps/$APP/helm-values-kgst.yaml
+  exit 0
+fi
